copy material effects from one material to another
this feature is needed at least at moment due to, when rendering a renderPass with a specific material, we need to copy all the effects values the renderPass material need from the original material in order it works. Another solution, could be to have more than one renderer per GameEntity and then, have a normal Renderer and the specific RenderPass Renderer with everything correctly set.