query working
The query for a region inside the quadtree is working.
Given a region, the query searches the min node which contains the region, and traverses the tree from that node to the leaves.
But, if an object is in a parent node of the node found, this object will not be returned as part of the results of the query.  And, it's an error.