givin support to the render pass to choice if use or not spacepartition
Now the rendererSystem is filled with all renderers in the gamescene.
The spacepartition only marks as visible those entities are inside the camera volume. And updates this mark every frame.
When a renderer is added to the RenderSystem, the renderPass accepts this renderer or not, depending on the layer mask and if accpets only space partition renderer those that are marked as visible.
In that way, each renderpass can choice if want to take in account all renderers or only those that are visible in the spacepartition.
The refraction, reflection and gameplay passes are marked as only spacepartition, and shadows not. (because we want shadows be visible for entities outside the spacepartition volume)