added hysteresis cicle to the LOD
Now, when a model changes its renderer by another depending on the distance of each renderer LOD, is taking into account an hysteresis cicle. That is, if the distance of the lod is D, is going to the next lod if d > D + hysteresisValue, and going down to the previous lod if d < D - hysteresisValue. In that way, the models stop blinking.