several fixes
Fixed refraction/reflection over water problem when particles_layer was setup in the refraction.
When refraction and reflection pass have the same layers, the system doesn't work well. Both renderpass need to have different layers number. I added a reflection_layer to fix this.
I also added clipping plane to particles, particles under water are not reflected.
And, added a new transparent layer to draw the grass with no collision with particles layer. If they are the same, its order is not ok because grass is using depth buffer particles not.