validation and removing code not needed
When creating render passes, we need to assure they not repeat a layer mask. If two render pass have the same layer mask when renderers were added into one of this render pass the render system will add twice, and will remain the second render pass empty. ValidateRenderPassesLayerMask tries to prevent it.
Then, we don't need to find a renderer in the list of rendereres per render pass due to not having render passes with the same layer mask, this possibility won't be possible.