pick up objects on the game by player
Added new parameter to know if light is enabled or not in the shaders
Added method into spacepartitionsystem to get gameentities near a point and radius
Added support to QuadTree to remove entities
Posponed deleting gameentities into the gamescene after a frame. The process is, add to notify to remove, notify all systems they remove its reference to the entity to remove and later on, remove the entities in the gamescene.
Added method to create a en entity from a model in the engine
ShooterGame create the first props
Created a Battery GameEntity
Created a PickeupComponent to indicate gameEntities can be piked up.