damage system ready
DamageSystem is ready and working
A GameEntity should have a HealthComponent a DamageComponent and a GameEventsComponent.
When the entity recieves damage through DamageController the DamageSystem updates its health, modifying the HealthComponent and puts a HealthEvent into the GameEventsComponent queue.
This way, the player can process the health event.
If the entity dies, a DieEvent will be enqueued (not yet)