Refactor
Due to a static function can't be virtual, we have implemented GameEntity as a template and then we can DoCreate related to the derived class.
Because of GameEntity had templated methods and now needed to be a templated class, we aggregate a IGameEntity with the templated methods, and GameEntity as a templated class. Seems a class cannot have both templates...