fix issue related to the debugsystem
If a GameEntity is disabled, won't be updated.
But, if we want to see bounding boxes, the disabled entity will show the bounding box.
The problem were, the debugSystem was updating the entities too, without taking into account if they were enabled or not. Anyway, the DebugSystem shouldn't update the entities.