fixed render problems
After adding particles to the scene, the renderer stop working as expected.
The problem was, the particles were not writing the depth buffer and the renderSystem was clearing buffer as it was enabled. The renderSystem should always enable depth writing at the beginning of render in order to glClear the depth buffer.
Similar problem happened with water, particles were not reflected. This is because the IFrameBuffer used in water and shadows where clearing the depth buffer to, without taking care if the last render call drop the depth buffer enabled or not.
Now is all fixed throught the RenderSystem which is the responsable to modify these values and remember the last ones.