fixed another problem due to particles
The problem was in the creation of the entities in the scene.
Build entity phase will build the entity itself,
When adding the already built entity into the list of entities in the scene, will be added first all of them and in a second phase, they will be initialized. That is because when initializing an entity this entity can clone existing entities.